refactor: 툴팁 내용 분리
