+++ apps/mac/src/renderer/src/features/home/ui/ranking/Ranking.constants.ts
@@ -1,0 +1,7 @@
+export const rankingRewardTooltipContent = `Aura: 100,000 EXP & 상위 1등
+Master: 75,000 EXP & 상위 3등
+Diamond: 50,000 EXP
+Gold: 30,000 EXP
+Silver: 10,000 EXP
+Bronze: 1,000 EXP
+Unranked: 1,000 EXP`;
