+++ apps/mac/src/renderer/src/widgets/topbar/Topbar.constants.ts
@@ -1,0 +1,21 @@
+export const expTooltipContent = `Github (1시간 간격으로 반영)
+
+커밋: 1개당 50점 (최대 50개)
+리뷰: 1개당 100점 (최대 5개)
+PR: 1개당 100점 (최대 5개)
+이슈: 1개당 10점 (최대 5개)
+
+학습 시간
+
+1분당 10점 (최대 10시간)
+오전 6시 갱신
+
+출석
+
+하루: 100점
+7일 연속 출석: 500 * 연속 출석 주
+
+로드맵
+
+챕터 클리어: 300점
+섹션 클리어: 2000점`;
